services/mcp/pyproject.toml
[project]
name = "story-sprout-mcp"
version = "0.2.0"
description = "FastMCP server for Story Sprout"
# readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.12.3",
    "mcp>=1.14.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.32.1",
]

[tool.setuptools]
packages = []

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.ruff]
line-length = 119
fix = true
lint.select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "I",   # isort
]
lint.pydocstyle.convention = "google"

[tool.ruff.format]
quote-style = "double"
skip-magic-trailing-comma = false
docstring-code-line-length = 79

[tool.pytest.ini_options]
python_files = ["tests.py", "test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-v",
]
asyncio_mode = "auto"
